+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Fixed
 * Optional constructor arguments are lost. [#2718](https://github.com/fsprojects/fantomas/issues/2718)
-* Empty nested module. [#2721](https://github.com/fsprojects/fantomas/issues/2721~~~~)
+* Bad format result with SynExpr.DotSet. [#2000](https://github.com/fsprojects/fantomas/issues/2000)
+* Empty nested module. [#2721](https://github.com/fsprojects/fantomas/issues/2721)
+* Optional parameter lost backticks. [#2731](https://github.com/fsprojects/fantomas/issues/2731)
 
 ## [5.2.0-alpha-012] - 2023-01-14