+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [5.2.0] - 2023-01-19
 
 ### Fixed
 * Optional constructor arguments are lost. [#2718](https://github.com/fsprojects/fantomas/issues/2718)